+++ Client/src/app/shared/components/locations-list/locations-list.component.ts
@@ -30,6 +30,8 @@ export class LocationsListComponent {
 
   editIndex = -1;
 
+  requesting = false;
+
   constructor(
     private store: Store<AppState>,
     private translate: TranslateService,
@@ -48,31 +50,43 @@ export class LocationsListComponent {
   }
 
   deleteItem(location: LocationInfo) {
-    this.translate.get('AreYouSureYouWantToDelete', { value: location.name }).subscribe((res: string) => {
-      const dialogData: SimpleDialogData = {
-        title: location.name,
-        body: res,
-        okButtonColor: 'warn',
-        okButtonLabel: this.labelOk,
-        cancelButtonLabel: this.labelCancel,
-      };
+    if (!this.requesting) {
+      this.requesting = true;
+      this.translate.get('AreYouSureYouWantToDeleteLocation', { value: location.name }).subscribe((res: string) => {
+        const dialogData: SimpleDialogData = {
+          title: location.name,
+          body: res,
+          okButtonColor: 'warn',
+          okButtonLabel: this.labelOk,
+          cancelButtonLabel: this.labelCancel,
+        };
 
-      // Open dialog for user to confirm action.
-      const dialogRef = this.dialog.open(SimpleDialogComponent, { data: dialogData });
-      dialogRef.afterClosed().subscribe((result) => {
-        if (result) {
-          this.locationService.deleteLocation(location.locationId).subscribe(
-            (data) => {
-              if (data) {
-                this.store.dispatch(this.locationActions.getMyLocations());
+        // Open dialog for user to confirm action.
+        const dialogRef = this.dialog.open(SimpleDialogComponent, { data: dialogData });
+        dialogRef.afterClosed().subscribe((result) => {
+          if (result) {
+            this.locationService.deleteLocation(location.locationId).subscribe(
+              (data) => {
+                if (data) {
+                  this.store.dispatch(this.locationActions.getMyLocations());
+                } else {
+                  this.showError();
+                }
+              },
+              () => {
+                this.showError();
               }
-            },
-            () => {
-              this.notificationService.showError(this.labelErrorOccurred, this.labelOk, 5000);
-            }
-          );
-        }
+            );
+          }
+        });
       });
-    });
+    }
+  }
+
+  showError() {
+    this.notificationService.showError(this.labelErrorOccurred, this.labelOk, 5000);
+    setTimeout(() => {
+      this.requesting = false;
+    }, 3000);
   }
 }
